pyproject.toml
[project]
name = "stl-explore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy-stl>=3.2.0",
    "openai>=1.79.0",
]
